refactor: useGetChannelName 훅 추가
