+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/_components/ChatDetailLayout/ChatDetailLayout.tsx
@@ -5,25 +5,14 @@ import { StrictPropsWithChildren } from '@/types/common';
 import Sidebar from '../Sidebar';
 import { useState } from 'react';
 import { MenuIcon } from '@/icons';
-import { useGetChannelName, useGetUsersInChannel } from '../../../_hooks/useQueryChat';
-import { getDmChannelName } from '../../../_utils/getDmChannelName';
-import { useWorkspaceUserId } from '@/hooks/useWorkspaceUserId';
-import { useParams } from 'next/navigation';
+import { useGetChannelName } from '../../_hook/useGetChannelName';
 
 const ChatDetailLayout = ({ children }: StrictPropsWithChildren) => {
   const [isOpenSidebar, setIsOpenSidebar] = useState(false);
-  const workspaceUserId = useWorkspaceUserId();
-  const { id } = useParams();
-  const stringId = Array.isArray(id) ? id[0] : id;
-  const { data: groupChannelName } = useGetChannelName({ id: stringId });
 
-  const { data: usersInChannel = {} } = useGetUsersInChannel({
-    channel_id: Number(stringId),
-    workspace_user_id: workspaceUserId
-  });
-
-  const dmChannelName = getDmChannelName(usersInChannel);
-  const channelName = dmChannelName || groupChannelName || '';
+  // TODO : 다른 페이지에서 이 정보를 가져올때 호출을 줄이는 것을 getQueryData로 해도 되나?, 그냥 호출해도 캐싱 되게 stale Time 해야하나?
+  // TODO : Context는 필요 없어 useQuery가 이미 그 역할을 해주니까..
+  const channelName = useGetChannelName();
 
   const handleOpenSidebar = () => {
     setIsOpenSidebar((prev) => !prev);
@@ -32,7 +21,7 @@ const ChatDetailLayout = ({ children }: StrictPropsWithChildren) => {
   return (
     <>
       <PageLayout
-        title={dmChannelName || channelName || ''}
+        title={channelName}
         showBottomBar={false}
         TopBarRightIcon1={<MenuButton onClick={handleOpenSidebar} aria-label="사이드바 열기" />}
         contentClassName="h-[calc(100dvh-52px)] overflow-hidden"
