+++ src/app/(providers)/(root)/[workspaceId]/chat/_utils/getChannelName.ts
